backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
# ML Dependencies
torch
transformers
peft
accelerate
protobuf
sentencepiece
sentencepiece
python-dotenv
groq
chromadb
sentence-transformers
pypdf
langchain-text-splitters
